+++ Shopify User/Modules/Products/View/ProductsViewController.swift
@@ -26,7 +26,8 @@ class ProductsViewController: UIViewController , UITableViewDelegate, UITableVie
     let dropDown = DropDown()
     var viewModel : GetProductsViewModel?
     var brandName : String?
-    let dataManager = DataManager.sharedInstance
+    let dataViewModel = ProductsDetailsViewModel(networkManager: NetworkManager(url: ""),dataManager: DataManager.sharedInstance)
+
     var customer_id = UserDefaults.standard.string(forKey: Constants.KEY_USER_ID)
     @IBOutlet weak var noItemFoundImg: UIImageView!
     @IBOutlet weak var priceSliderOutlet: UISlider!
@@ -104,7 +105,7 @@ class ProductsViewController: UIViewController , UITableViewDelegate, UITableVie
         
         let reachability = try! Reachability()
         if reachability.connection != .unavailable{
-            viewModel = GetProductsViewModel(dataManager: dataManager)
+            viewModel = GetProductsViewModel()
             viewModel?.brandName = brandName
             
             viewModel?.bindResultToViewController={
@@ -150,7 +151,7 @@ class ProductsViewController: UIViewController , UITableViewDelegate, UITableVie
     func colorHeart(product:Product , favBtn : UIButton){
         let coreData = ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: customer_id)
         //print("My p  :\(product_VC)")
-        let is_Exist = dataManager.isProductExist(myProduct: coreData)
+        let is_Exist = dataViewModel.isExistIntoDB(product: coreData)
         if(is_Exist){
             favBtn.imageView?.image = UIImage(systemName: "heart.fill")
             print("product already saved")
@@ -174,7 +175,7 @@ class ProductsViewController: UIViewController , UITableViewDelegate, UITableVie
          let product = self.productsList[indexPath.row]
         colorHeart(product: product, favBtn: cell.favBtn)
          productCoreData = ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: self.customer_id)
-        let is_Exist = dataManager.isProductExist(myProduct: productCoreData)
+        let is_Exist = dataViewModel.isExistIntoDB(product: productCoreData)
         cell.setFavUI(isFav: is_Exist)
         let url = URL(string: productsList[indexPath.row].image?.src ?? "")
         
@@ -228,29 +229,70 @@ class ProductsViewController: UIViewController , UITableViewDelegate, UITableVie
     
     func clicked(_ row: Int) {
         //print("My p  :\(product_VC)")
-        let product = productsList[row]
-        productCoreData = ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: self.customer_id)
-        
-         let is_Exist = dataManager.isProductExist(myProduct: productCoreData)
-         if is_Exist{
-            let alert : UIAlertController = UIAlertController(title: "ALERT!", message: "ARE YOU SURE TO DELETE FROM FAVORITE?", preferredStyle: .alert)
+        guard let state = UserDefaults.standard.string(forKey: Constants.KEY_USER_STATE) else{return}
+        if(state == Constants.USER_STATE_GUEST){
+            AlertCreator.SignUpAlert(viewController: self)
+        } else {
+            let product = productsList[row]
+            productCoreData = ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: self.customer_id)
             
-            alert.addAction(UIAlertAction(title: "YES", style: .default,handler: { [weak self] action in
-//                    delete fromm fav
-                self?.viewModel?.deleteFromDB(product: ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: self?.customer_id))
-                self?.tableView.reloadData()
+            let is_Exist = dataViewModel.isExistIntoDB(product: productCoreData)
+            if is_Exist{
+                let alert : UIAlertController = UIAlertController(title: "ALERT!", message: "ARE YOU SURE TO DELETE FROM FAVORITE?", preferredStyle: .alert)
                 
-            }))
-             alert.addAction(UIAlertAction(title: "NO", style: .cancel,handler: { [weak self] action in
-                 self?.tableView.reloadData()
-             }))
-            present(alert, animated: true, completion: nil)
-        }
-        else{
-//             insert to fav
-            viewModel?.insertIntoDB(product: ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: customer_id))
-            tableView.reloadData()
-        }
+                alert.addAction(UIAlertAction(title: "YES", style: .default,handler: { [weak self] action in
+                    //                    delete fromm fav
+                    self?.dataViewModel.deleteFromDB(product: ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: self?.customer_id))
+                    self?.tableView.reloadData()
+                    
+                    self?.dataViewModel.bindCartData = { [weak self] in
+                        var myDraft = self?.dataViewModel.wishListArray
+                        var arr = myDraft?.line_items
+                      //  print("Array before deletion \(arr)")
+                        var myId = product.id
+                      //  print("myId = \(myId)")
+
+                        for i in 0..<(arr?.count ?? 0) {
+                            var apiId = arr?[i].product_id
+                          //  print("apiId = \(apiId)")
+                            if (apiId == myId){
+                                arr?.remove(at: i)
+                                self?.dataViewModel.deleteFromDB(product:ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: self?.customer_id) )
+                                break
+                            }
+                        }
+                      //  print("Array after deletion \(arr)")
+
+                        myDraft?.line_items? = arr ?? []
+                        self?.dataViewModel.updateWishList(wishListItem: myDraft ?? Draft_orders())
+                    }
+                    self?.dataViewModel.loadWishListItems()
+                    
+                }))
+                alert.addAction(UIAlertAction(title: "NO", style: .cancel,handler: { [weak self] action in
+                    self?.tableView.reloadData()
+                }))
+                present(alert, animated: true, completion: nil)
+            }
+            else{
+                //             insert to fav
+//                dataViewModel.insertIntoDB(product:)
+//                tableView.reloadData()
+                
+                dataViewModel.insertIntoDB(product:  ProductCoreData(id: product.id,title: product.title,price: product.variants?[0].price,Pimage: product.image?.src,user_id: customer_id))
+               // favBtn.setImage(UIImage(systemName: "heart.fill"), for: .normal)
+                dataViewModel.bindCartData = { [weak self] in
+                    var myDraft = self?.dataViewModel.wishListArray
+                    var arr = myDraft?.line_items
+                    let pro = Constants().mapProductToLineItems(product: product )
+                    arr?.append(pro)
+                    myDraft?.line_items? = arr ?? []
+                    self?.dataViewModel.updateWishList(wishListItem: myDraft ?? Draft_orders())
+                }
+                dataViewModel.loadWishListItems()
+                tableView.reloadData()
+            }
+       }
     }
 }
